requirements.txt
tensorflow
numpy
pandas
matplotlib
seaborn
streamlit
streamlit-drawable-canvas
pillow
